+++ src/nrepl/connectSequence.ts
@@ -9,6 +9,7 @@ enum ProjectTypes {
     "shadow-cljs" = "shadow-cljs",
     "lein-shadow" = "lein-shadow",
     "babashka" = "babashka",
+    "nbb" = "nbb",
     'generic' = 'generic',
     'cljs-only' = 'cljs-only'
 }
@@ -153,7 +154,14 @@ const babashkaDefaults: ReplConnectSequence[] = [{
     name: "Babashka",
     projectType: ProjectTypes['babashka'],
     cljsType: CljsTypes.none,
-    nReplPortFile: ["bb-nrepl.port"]
+    nReplPortFile: [".bb-nrepl.port"]
+}];
+
+const nbbDefaults: ReplConnectSequence[] = [{
+    name: "nbb",
+    projectType: ProjectTypes['nbb'],
+    cljsType: CljsTypes["ClojureScript nREPL"],
+    nReplPortFile: [".nbb-nrepl.port"]
 }];
 
 const defaultSequences = {
@@ -163,6 +171,7 @@ const defaultSequences = {
     "lein-shadow": leinShadowDefaults,
     'generic': genericDefaults,
     'babashka': babashkaDefaults,
+    'nbb': nbbDefaults,
     'cljs-only': cljsOnlyDefaults
 };